+++ apps/explorer/lib/explorer/chain/import/runner/internal_transactions.ex
@@ -271,7 +271,7 @@ defmodule Explorer.Chain.Import.Runner.InternalTransactions do
         where: t.block_hash in ^pending_block_hashes,
         select: map(t, [:hash, :block_hash, :block_number, :cumulative_gas_used]),
         # Enforce Transaction ShareLocks order (see docs: sharelocks.md)
-        order_by: t.hash,
+        order_by: [asc: t.hash],
         lock: "FOR UPDATE"
       )